Removed user-space caching of open chdb files.
While this modification can make some artificial benchmarks slower, it
is unlikely that it will significantly affect any realistic workload,
since the time it takes to load and unload a chdb file is in the order
of 10 microseconds.

On the other hand by avoiding user-space caching, the newest version of
a file can be loaded each time a PHP chdb instance is created, thus
allowing the use of chdb for slowly changing data. Chdb files are now
also unloaded as soon as the corresponding chdb object is destroyed,
making better use of memory in certain scenarios.

Kernel-level caching should still provide reuse of the same physical
pages across multiple requests.