Bug 22600: fix possible segfault in 64 bit systems
Because of a missing include the compiler was implicitly declaring the
PHP functions for throwing exceptions. This could segfault in a 64-bit
system because the pointer returned from zend_exception_get_default was
being truncated to a 32-bit value.

This fixes the problem, enables all compiler warnings, and cleans-up
some other code a bit.