Use RuntimeException instead of Exception for runtime errors
As reported at https://github.com/lcastelli/chdb/issues/2 - PHP
provides a more specific Exception class in the SPL extension for
runtime errors. The use of this type is thus more precise.

This change makes CHDB dependent on the SPL extension, but since it's
compiled by default since PHP 5.0 and required since PHP 5.3 this
shouldn't be a big issue.